+++ activity/src/main/java/com/jjg/game/activity/firstpayment/message/bean/FirstPaymentDetailInfo.java
@@ -17,9 +17,9 @@ public class FirstPaymentDetailInfo extends BaseActivityDetailInfo {
     @ProtoDesc("购买奖励")
     public List<ItemInfo> bugGet;
     @ProtoDesc("需要充值金额")
-    public int rechargePrice;
+    public double rechargePrice;
     @ProtoDesc("原价值")
-    public int wasPrice;
+    public double wasPrice;
     @ProtoDesc("超值")
-    public  int bestValue;
+    public double bestValue;
 }
