+++ hall/src/main/java/com/jjg/game/hall/pointsaward/PointsAwardService.java
@@ -330,7 +330,7 @@ public class PointsAwardService {
      */
     public void recharge(Order order) {
         long playerId = order.getPlayerId();
-        long price = order.getPrice();
+        long price = order.getPrice().longValue();
         RLock lock = rechargeMap.getReadWriteLock(playerId).writeLock();
         if (lock.tryLock()) {
             long resultValue = getRecharge(playerId) + price;
